chore: move erc20 proposal function to erc20 file
