+++ tests/e2e/e2e_setup_test.go
@@ -14,7 +14,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
 	"github.com/spf13/viper"
@@ -784,29 +783,6 @@ func (s *IntegrationTestSuite) writeExpeditedSoftwareUpgradeProp(c *chain) {
 }`
 
 	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalExpeditedSoftwareUpgrade), []byte(body))
-	s.Require().NoError(err)
-}
-
-func (s *IntegrationTestSuite) writeERC20RegisterProposal(c *chain, erc20Address common.Address) {
-	body := `{
-		"messages": [
-		 {
-		  "@type": "/cosmos.evm.erc20.v1.MsgRegisterERC20",
-		  "authority": "kii10d07y265gmmuvt4z0w9aw880jnsr700jrff0qv",
-		  "erc20addresses": [
-		    "%s"
-		  ]
-		 }
-		],
-		"metadata": "ipfs://CID",
-		"deposit": "100akii",
-		"title": "title",
-		"summary": "test"
-	   }`
-
-	propMsgBody := fmt.Sprintf(body, erc20Address.String())
-
-	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalRegisterERC20), []byte(propMsgBody))
 	s.Require().NoError(err)
 }
